* commit '91a3a7b5bb9145860ab20c13a7e250e33493340c':
  Review compiler warnings.
  Refactor removing deprecated usages.

+++ osx/src/main/objc/CDBookmarkCell.m
@@ -101,7 +101,7 @@
 	}
 
     BOOL highlighted = [self isHighlighted] && ![[self highlightColorWithFrame:cellFrame inView:controlView] isEqualTo:[NSColor secondarySelectedControlColor]];
-    int size = [[NSUserDefaults standardUserDefaults] integerForKey:@"bookmark.icon.size"];
+    NSInteger size = [[NSUserDefaults standardUserDefaults] integerForKey:@"bookmark.icon.size"];
 	
     NSDictionary *nicknameFont;
     if(LARGE_BOOKMARK_SIZE == size) {
